chore(pkg/utils): Export cli utils for re-use
