+++ pkg/metricsstore/metricsstore_test.go
@@ -7,6 +7,8 @@ import (
 	"testing"
 
 	tassert "github.com/stretchr/testify/assert"
+
+	"github.com/openservicemesh/osm/pkg/constants"
 )
 
 func setup() {
@@ -39,7 +41,7 @@ func TestMetricsStore(t *testing.T) {
 
 			handler := DefaultMetricsStore.Handler()
 
-			req, err := http.NewRequest("GET", "/metrics", nil)
+			req, err := http.NewRequest("GET", constants.HTTPServerMetricsPath, nil)
 			assert.Nil(err)
 
 			rr := httptest.NewRecorder()
@@ -95,7 +97,7 @@ osm_proxy_connect_count %d
 
 			handler := DefaultMetricsStore.Handler()
 
-			req, err := http.NewRequest("GET", "/metrics", nil)
+			req, err := http.NewRequest("GET", constants.HTTPServerMetricsPath, nil)
 			assert.Nil(err)
 
 			rr := httptest.NewRecorder()
